backend/app/services/hippostcard_service.py
import os
import httpx
import re
from typing import List, Dict, Any, Optional
from bs4 import BeautifulSoup
from dotenv import load_dotenv
from app.models.search_models import SearchResult, SearchFilters

# Load environment variables
load_dotenv()

# HipPostcard search URL
HIPPOSTCARD_SEARCH_URL = "https://www.hippostcard.com/search"
HIPPOSTCARD_AFFILIATE_ID = os.getenv("HIPPOSTCARD_AFFILIATE_ID", "")  # Affiliate ID if available

# Use mock data for development
USE_MOCK_DATA = False  # Set to False for production

async def search_hippostcard(query: str, filters: Optional[SearchFilters] = None, page: int = 1, limit: int = 20) -> List[SearchResult]:
    """
    Search for postcards on HipPostcard.
    
    Note: This is a simplified implementation that scrapes the HipPostcard search results page.
    In a production environment, you would typically use their official API if available.
    
    Args:
        query: The search query
        filters: Optional search filters
        page: Page number for pagination
        limit: Number of results per page
        
    Returns:
        List of SearchResult objects
    """
    # Use mock data only if explicitly enabled
    if USE_MOCK_DATA:
        return await search_hippostcard_mock(query, filters, page, limit)
    
    try:
        # Prepare query parameters
        params = {
            "keywords": query,
            "page": page
        }
        
        # Add price filters if provided
        if filters:
            if filters.price_min is not None:
                params["min_price"] = filters.price_min
            if filters.price_max is not None:
                params["max_price"] = filters.price_max
        
        # Make HTTP request to HipPostcard search page
        async with httpx.AsyncClient() as client:
            headers = {
                "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36"
            }
            
            response = await client.get(HIPPOSTCARD_SEARCH_URL, params=params, headers=headers)
            
            if response.status_code != 200:
                print(f"HipPostcard search failed: {response.status_code}")
                return []
            
            # Parse HTML response
            soup = BeautifulSoup(response.text, "html.parser")
            
            # Find postcard listings
            # Note: This is a placeholder implementation and would need to be updated
            # based on the actual HTML structure of HipPostcard's search results page
            listing_elements = soup.select(".postcard-item")  # Update selector based on actual HTML
            
            results = []
            for i, element in enumerate(listing_elements):
                if i >= limit:
                    break
                
                try:
                    # Extract data from HTML elements
                    # These selectors would need to be updated based on actual HTML structure
                    title_element = element.select_one(".postcard-title")
                    price_element = element.select_one(".postcard-price")
                    image_element = element.select_one(".postcard-image img")
                    link_element = element.select_one("a.postcard-link")
                    
                    title = title_element.text.strip() if title_element else "Untitled Postcard"
                    
                    # Extract price
                    price = 0.0
                    currency = "USD"
                    if price_element:
                        price_text = price_element.text.strip()
                        price_match = re.search(r'(\d+\.\d+)', price_text)
                        if price_match:
                            price = float(price_match.group(1))
                    
                    # Get image URL
                    image_url = ""
                    if image_element and image_element.has_attr("src"):
                        image_url = image_element["src"]
                    
                    # Get listing URL
                    link = ""
                    if link_element and link_element.has_attr("href"):
                        link = link_element["href"]
                        if not link.startswith("http"):
                            link = f"https://www.hippostcard.com{link}"
                    
                    # Create affiliate link if affiliate ID is available
                    affiliate_link = None
                    if HIPPOSTCARD_AFFILIATE_ID and link:
                        affiliate_link = f"{link}?ref={HIPPOSTCARD_AFFILIATE_ID}"
                    
                    # Extract date and location from title if available
                    date = None
                    location = None
                    
                    # Simple extraction - in a real app, use more sophisticated NLP
                    year_match = re.search(r'(18|19|20)\d{2}', title)
                    if year_match:
                        date = year_match.group(0)
                    
                    # Create SearchResult
                    result = SearchResult(
                        source="HipPostcard",
                        title=title,
                        image_url=image_url,
                        price=price,
                        currency=currency,
                        link=link,
                        description="",  # No description available in search results
                        date=date,
                        location=location,
                        affiliate_link=affiliate_link
                    )
                    
                    results.append(result)
                
                except Exception as e:
                    print(f"Error parsing HipPostcard listing: {str(e)}")
                    continue
            
            return results
    
    except Exception as e:
        print(f"HipPostcard search error: {str(e)}")
        return []

# Alternative implementation using a mock API response for development
async def search_hippostcard_mock(query: str, filters: Optional[SearchFilters] = None, page: int = 1, limit: int = 20) -> List[SearchResult]:
    """
    Mock implementation of HipPostcard search for development purposes.
    
    Args:
        query: The search query
        filters: Optional search filters
        page: Page number for pagination
        limit: Number of results per page
        
    Returns:
        List of SearchResult objects with mock data
    """
    # Generate mock results based on query
    results = []
    for i in range(min(limit, 10)):  # Generate up to 10 mock results
        # Create a mock title based on the query
        title = f"Vintage Postcard - {query.title()} - {1900 + i*10}"
        
        # Create SearchResult with mock data
        result = SearchResult(
            source="HipPostcard",
            title=title,
            image_url=f"https://example.com/postcard{i}.jpg",
            price=5.99 + i,
            currency="USD",
            link=f"https://www.hippostcard.com/listing/{i}",
            description=f"Beautiful vintage postcard featuring {query}. Circa {1900 + i*10}.",
            date=str(1900 + i*10),
            location=None,
            affiliate_link=None
        )
        
        results.append(result)
    
    return results 