backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.6.3
openai==1.27.0
python-multipart==0.0.9
pytesseract==0.3.10
Pillow==10.2.0
mangum==0.17.0
starlette==0.36.3
jinja2==3.1.3 